Simple fix for strict validation
**Why?**
We have switched to marshmallow 3, which uses strict schema validation. This means that any unexpected field sent in a request triggers a `422`.
`microcosm-resourcesync` assumes that it can take a resource and send all of it to a new endpoint, which doesn't work with strict validation.

**What?**

Filter out `id` and `_links` fields before sending the resource. This should work with a YAML file source, and won't work in general with an HTTP source. In practice though, we only use `microcosm-resourcesync` in two places (see https://github.com/globality-corp/miranda/blob/develop/miranda/data/cli/synchronize.py and https://github.com/globality-corp/blanca/blob/develop/blanca/data/main.py), and both use a YAML file as a source.